feature: Add WolframSession::{enter_and_wait_with_output_handler, try_enter_and_wait_with_output_handler}
These new evaluation methods take an `&mut dyn FnMut(Output)` handler
function that is called immediately when printed output is sent from
the Kernel to the client.

This makes it easy to perform blocking evaluations that still print
output as it comes in, which is especially important for long-running
evaluations that print progress updates, or which may generate errors
indicating to the user they may want to abort the evaluation instead
of waiting for it to finish.